Convert items hooks to helper methods
