+++ src/hooks/items.js
@@ -29,50 +29,6 @@ export const getFilteredLiveCheckOffsData = (applyFilter = a => a, forTeacher = 
   return checkOffs;
 };
 
-// ******************
-// TODO: Remove these hooks and replace them with direct helper function calls in the components.
-// ******************
-export const useUnlockedItems = (items, locked, whiteList) =>
-  locked ? useMemo(() => filterPages(whiteList, items), [items, whiteList]) : items;
-
-export const useFlatItems = items => useMemo(() => flattenPages(items), [items]);
-
-export const useFlatUnlockedItems = (items, locked, whiteList) =>
-  useFlatItems(useUnlockedItems(items, locked, whiteList));
-
-/* ===============================
- * === Custom Helper Functions ===
- * =============================== */
-
-/**
- * Check the given `value` object against the given `unlocked` whitelist.
- * Return only the allowed values.
- */
-export const filterPages = (unlocked, value, prefix = '') => {
-  const newValue = {};
-  Object.entries(value).forEach(([key, val]) => {
-    if (typeof val === 'string' && unlocked.includes(prefix + key)) newValue[key] = val;
-    else if (typeof val !== 'string') {
-      const nextUp = filterPages(unlocked, val, `${prefix}${key}.`);
-      if (Object.keys(nextUp).length > 0) newValue[key] = nextUp;
-    }
-  });
-  return newValue;
-};
-
-/**
- * Flatten a multi-dimentional object into a single array of pages.
- */
-export const flattenPages = (fItems, prefix = '') => {
-  const flattened = [];
-  Object.entries(fItems).forEach(([key, value]) => {
-    const item = prefix + key;
-    if (typeof value === 'string') flattened.push(item);
-    else flattened.push(...flattenPages(value, `${item}.`));
-  });
-  return flattened;
-};
-
 /* ===============================
  * === Custom Reusable Effects ===
  * =============================== */
